requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.3.0
corsheaders==4.9.0
stripe==10.12.0
requests==2.32.3
django-filter==24.2.0
drf-yasg==1.21.6